+++ src/app/core/core.module.ts
@@ -7,6 +7,7 @@ import { RouterModule } from '@angular/router';
 import { NotifyModule } from './notify/notify.module';
 import { LocalBackupModule } from '../imex/local-backup/local-backup.module';
 import { BannerModule } from './banner/banner.module';
+import {CompressionModule} from './compression/compression.module';
 
 @NgModule({
   imports: [
@@ -18,6 +19,7 @@ import { BannerModule } from './banner/banner.module';
     BannerModule,
     NotifyModule,
     LocalBackupModule,
+    CompressionModule,
   ],
   exports: [
     PersistenceModule,