+++ src/app/core/persistence/persistence.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 import {
   LS_BACKUP,
   LS_BOOKMARK_STATE,
@@ -6,6 +6,7 @@ import {
   LS_ISSUE_STATE,
   LS_LAST_ACTIVE,
   LS_NOTE_STATE,
+  LS_PROJECT_ARCHIVE,
   LS_PROJECT_META_LIST,
   LS_PROJECT_PREFIX,
   LS_REMINDER,
@@ -13,22 +14,27 @@ import {
   LS_TASK_ATTACHMENT_STATE,
   LS_TASK_STATE
 } from './ls-keys.const';
-import { GlobalConfig } from '../../features/config/config.model';
-import { IssueProviderKey, IssueState } from '../../features/issue/issue';
-import { ProjectState } from '../../features/project/store/project.reducer';
-import { TaskState } from '../../features/tasks/store/task.reducer';
-import { EntityState } from '@ngrx/entity';
-import { Task, TaskWithSubTasks } from '../../features/tasks/task.model';
-import { AppDataComplete } from '../../imex/sync/sync.model';
-import { BookmarkState } from '../../features/bookmark/store/bookmark.reducer';
-import { AttachmentState } from '../../features/attachment/store/attachment.reducer';
-import { NoteState } from '../../features/note/store/note.reducer';
-import { Reminder } from '../../features/reminder/reminder.model';
-import { SnackService } from '../snack/snack.service';
-import { DatabaseService } from './database.service';
-import { loadFromLs, saveToLs } from './local-storage';
-import { issueProviderKeys } from '../../features/issue/issue.const';
-import { DEFAULT_PROJECT_ID } from '../../features/project/project.const';
+import {GlobalConfig} from '../../features/config/config.model';
+import {IssueProviderKey, IssueState, IssueStateMap} from '../../features/issue/issue';
+import {ProjectState} from '../../features/project/store/project.reducer';
+import {TaskState} from '../../features/tasks/store/task.reducer';
+import {EntityState} from '@ngrx/entity';
+import {Task, TaskWithSubTasks} from '../../features/tasks/task.model';
+import {AppDataComplete} from '../../imex/sync/sync.model';
+import {BookmarkState} from '../../features/bookmark/store/bookmark.reducer';
+import {AttachmentState} from '../../features/attachment/store/attachment.reducer';
+import {NoteState} from '../../features/note/store/note.reducer';
+import {Reminder} from '../../features/reminder/reminder.model';
+import {SnackService} from '../snack/snack.service';
+import {DatabaseService} from './database.service';
+import {loadFromLs, saveToLs} from './local-storage';
+import {GITHUB_TYPE, issueProviderKeys, JIRA_TYPE} from '../../features/issue/issue.const';
+import {DEFAULT_PROJECT_ID} from '../../features/project/project.const';
+import {ArchivedProject, ProjectArchive} from '../../features/project/project.model';
+import {JiraIssueState} from '../../features/issue/jira/jira-issue/store/jira-issue.reducer';
+import {GithubIssueState} from '../../features/issue/github/github-issue/store/github-issue.reducer';
+import {CompressionService} from '../compression/compression.service';
+
 
 @Injectable({
   providedIn: 'root',
@@ -39,6 +45,7 @@ export class PersistenceService {
   constructor(
     private _snackService: SnackService,
     private _databaseService: DatabaseService,
+    private _compressionService: CompressionService,
   ) {
   }
 
@@ -60,6 +67,7 @@ export class PersistenceService {
     return this._saveToDb(LS_REMINDER, reminders, isForce);
   }
 
+  // TASKS
   async saveTasksForProject(projectId, taskState: TaskState, isForce = false): Promise<any> {
     return this._saveToDb(this._makeProjectKey(projectId, LS_TASK_STATE), taskState, isForce);
   }
@@ -68,6 +76,11 @@ export class PersistenceService {
     return this._loadFromDb(this._makeProjectKey(projectId, LS_TASK_STATE));
   }
 
+  async removeTasksForProject(projectId): Promise<TaskState> {
+    return this._removeFromDb(this._makeProjectKey(projectId, LS_TASK_STATE));
+  }
+
+  // TASK ARCHIVE
   async saveToTaskArchiveForProject(projectId, tasksToArchive: EntityState<TaskWithSubTasks>, isForce = false) {
     const lsKey = this._makeProjectKey(projectId, LS_TASK_ARCHIVE);
     const currentArchive: EntityState<Task> = await this._loadFromDb(lsKey);
@@ -91,6 +104,10 @@ export class PersistenceService {
     return this._loadFromDb(this._makeProjectKey(projectId, LS_TASK_ARCHIVE));
   }
 
+  async removeTaskArchiveForProject(projectId: string): Promise<EntityState<TaskWithSubTasks>> {
+    return this._removeFromDb(this._makeProjectKey(projectId, LS_TASK_ARCHIVE));
+  }
+
   async removeTasksFromArchive(projectId: string, taskIds: string[]) {
     const lsKey = this._makeProjectKey(projectId, LS_TASK_ARCHIVE);
     const currentArchive: EntityState<Task> = await this._loadFromDb(lsKey);
@@ -109,6 +126,7 @@ export class PersistenceService {
     }, true);
   }
 
+  // ISSUES
   async saveIssuesForProject(projectId, issueType: IssueProviderKey, data: IssueState, isForce = false): Promise<any> {
     return this._saveToDb(this._makeProjectKey(projectId, LS_ISSUE_STATE, issueType), data, isForce);
   }
@@ -117,6 +135,11 @@ export class PersistenceService {
     return this._loadFromDb(this._makeProjectKey(projectId, LS_ISSUE_STATE, issueType));
   }
 
+  async removeIssuesForProject(projectId, issueType: IssueProviderKey): Promise<IssueState> {
+    return this._removeFromDb(this._makeProjectKey(projectId, LS_ISSUE_STATE, issueType));
+  }
+
+  // BOOKMARKS
   async saveBookmarksForProject(projectId, bookmarkState: BookmarkState, isForce = false) {
     return this._saveToDb(this._makeProjectKey(projectId, LS_BOOKMARK_STATE), bookmarkState, isForce);
   }
@@ -125,6 +148,11 @@ export class PersistenceService {
     return this._loadFromDb(this._makeProjectKey(projectId, LS_BOOKMARK_STATE));
   }
 
+  async removeBookmarksForProject(projectId): Promise<BookmarkState> {
+    return this._removeFromDb(this._makeProjectKey(projectId, LS_BOOKMARK_STATE));
+  }
+
+  // ATTACHMENTS
   async saveTaskAttachmentsForProject(projectId, attachmentState: AttachmentState, isForce = false) {
     return this._saveToDb(this._makeProjectKey(projectId, LS_TASK_ATTACHMENT_STATE), attachmentState, isForce);
   }
@@ -133,12 +161,21 @@ export class PersistenceService {
     return this._loadFromDb(this._makeProjectKey(projectId, LS_TASK_ATTACHMENT_STATE));
   }
 
+  async removeTaskAttachmentsForProject(projectId): Promise<AttachmentState> {
+    return this._removeFromDb(this._makeProjectKey(projectId, LS_TASK_ATTACHMENT_STATE));
+  }
+
+  // NOTES
   async saveNotesForProject(projectId, noteState: NoteState, isForce = false) {
     return this._saveToDb(this._makeProjectKey(projectId, LS_NOTE_STATE), noteState, isForce);
   }
 
   async loadNotesForProject(projectId): Promise<NoteState> {
     return this._loadFromDb(this._makeProjectKey(projectId, LS_NOTE_STATE));
+  }
+
+  async removeNotesForProject(projectId): Promise<NoteState> {
+    return this._removeFromDb(this._makeProjectKey(projectId, LS_NOTE_STATE));
   }
 
   // GLOBAL CONFIG
@@ -149,6 +186,92 @@ export class PersistenceService {
 
   async saveGlobalConfig(globalConfig: GlobalConfig, isForce = false) {
     return this._saveToDb(LS_GLOBAL_CFG, globalConfig, isForce);
+  }
+
+  // PROJECT ARCHIVING
+  // -----------------
+  async loadProjectArchive(): Promise<ProjectArchive> {
+    return await this._loadFromDb(LS_PROJECT_ARCHIVE);
+  }
+
+  async saveProjectArchive(data: ProjectArchive): Promise<any> {
+    return await this._saveToDb(LS_PROJECT_ARCHIVE, data);
+  }
+
+  async loadArchivedProject(projectId): Promise<ArchivedProject> {
+    const archive = await this._loadFromDb(LS_PROJECT_ARCHIVE);
+    const projectDataCompressed = archive[projectId];
+    const decompressed = await this._compressionService.decompress(projectDataCompressed);
+    const parsed = JSON.parse(decompressed);
+    console.log(`Decompressed project, size before: ${projectDataCompressed.length}, size after: ${decompressed.length}`, parsed);
+    return parsed;
+  }
+
+  async removeArchivedProject(projectId): Promise<any> {
+    const archive = await this._loadFromDb(LS_PROJECT_ARCHIVE);
+    delete archive[projectId];
+    this.saveProjectArchive(archive);
+  }
+
+  async saveArchivedProject(projectId, archivedProject: ArchivedProject) {
+    const current = await this.loadProjectArchive() || {};
+    const jsonStr = JSON.stringify(archivedProject);
+    const compressedData = await this._compressionService.compress(jsonStr);
+    console.log(`Compressed project, size before: ${jsonStr.length}, size after: ${compressedData.length}`, archivedProject);
+    return this.saveProjectArchive({
+      ...current,
+      [projectId]: compressedData,
+    });
+  }
+
+  async loadCompleteForProject(projectId: string): Promise<ArchivedProject> {
+    const issueStateMap: IssueStateMap = {
+      JIRA: await this.loadIssuesForProject(projectId, JIRA_TYPE) as JiraIssueState,
+      GITHUB: await this.loadIssuesForProject(projectId, GITHUB_TYPE) as GithubIssueState,
+    };
+
+    return {
+      note: await this.loadNotesForProject(projectId),
+      bookmark: await this.loadBookmarksForProject(projectId),
+      task: await this.loadTasksForProject(projectId),
+      taskArchive: await this.loadTaskArchiveForProject(projectId),
+      taskAttachment: await this.loadTaskAttachmentsForProject(projectId),
+      issue: issueStateMap,
+    };
+  }
+
+  async removeCompleteRelatedDataForProject(projectId: string): Promise<any> {
+    await this.removeTasksForProject(projectId);
+    await this.removeTaskArchiveForProject(projectId);
+    await this.removeNotesForProject(projectId);
+    await this.removeTaskAttachmentsForProject(projectId);
+    await this.removeBookmarksForProject(projectId);
+    await issueProviderKeys.forEach(async (key) => {
+      await this.removeIssuesForProject(projectId, key);
+    });
+  }
+
+  async restoreCompleteRelatedDataForProject(projectId: string, data: ArchivedProject): Promise<any> {
+    await this.saveTasksForProject(projectId, data.task);
+    await this.saveToTaskArchiveForProject(projectId, data.taskArchive);
+    await this.saveNotesForProject(projectId, data.note);
+    await this.saveTaskAttachmentsForProject(projectId, data.taskAttachment as AttachmentState);
+    await this.saveBookmarksForProject(projectId, data.bookmark);
+    await issueProviderKeys.forEach(async (key) => {
+      await this.saveIssuesForProject(projectId, key, data.issue[key]);
+    });
+  }
+
+  async archiveProject(projectId: string): Promise<any> {
+    const projectData = await this.loadCompleteForProject(projectId);
+    await this.saveArchivedProject(projectId, projectData);
+    await this.removeCompleteRelatedDataForProject(projectId);
+  }
+
+  async unarchiveProject(projectId: string): Promise<any> {
+    const projectData = await this.loadArchivedProject(projectId);
+    await this.restoreCompleteRelatedDataForProject(projectId, projectData);
+    await this.removeArchivedProject(projectId);
   }
 
   // BACKUP AND SYNC RELATED
@@ -182,6 +305,7 @@ export class PersistenceService {
       project: await this.loadProjectsMeta(),
       globalConfig: await this.loadGlobalConfig(),
       reminders: await this.loadReminders(),
+      archivedProjects: await this.loadProjectArchive(),
       bookmark: await this._loadForProjectIds(pids, this.loadBookmarksForProject.bind(this)),
       note: await this._loadForProjectIds(pids, this.loadNotesForProject.bind(this)),
       task: await this._loadForProjectIds(pids, this.loadTasksForProject.bind(this)),
@@ -196,7 +320,7 @@ export class PersistenceService {
 
         return {
           ...prevAcc,
-          [projectId]: issueStateMap
+          [projectId]: issueStateMap as IssueStateMap
         };
       }, Promise.resolve({})),
     };
@@ -219,6 +343,7 @@ export class PersistenceService {
       this.saveProjectsMeta(data.project, true),
       this.saveGlobalConfig(data.globalConfig, true),
       this.saveReminders(data.reminders, true),
+      this.saveProjectArchive(data.archivedProjects),
       this._saveForProjectIds(data.bookmark, this.saveBookmarksForProject.bind(this), true),
       this._saveForProjectIds(data.note, this.saveNotesForProject.bind(this), true),
       this._saveForProjectIds(data.task, this.saveTasksForProject.bind(this), true),
@@ -267,6 +392,15 @@ export class PersistenceService {
     } else {
       console.warn('BLOCKED SAVING for ', key);
       return Promise.reject('Data import currently in progress. Saving disabled');
+    }
+  }
+
+  private async _removeFromDb(key: string, isForce = false): Promise<any> {
+    if (!this._isBlockSaving || isForce === true) {
+      return this._databaseService.remove(key);
+    } else {
+      console.warn('BLOCKED SAVING for ', key);
+      return Promise.reject('Data import currently in progress. Removing disabled');
     }
   }