+++ src/app/features/google/google-api.service.ts
@@ -1,22 +1,22 @@
-import { Injectable } from '@angular/core';
-import { GOOGLE_DEFAULT_FIELDS_FOR_DRIVE, GOOGLE_DISCOVERY_DOCS, GOOGLE_SCOPES, GOOGLE_SETTINGS } from './google.const';
+import {Injectable} from '@angular/core';
+import {GOOGLE_DEFAULT_FIELDS_FOR_DRIVE, GOOGLE_DISCOVERY_DOCS, GOOGLE_SCOPES, GOOGLE_SETTINGS} from './google.const';
 import * as moment from 'moment-mini';
-import { IS_ELECTRON } from '../../app.constants';
-import { MultiPartBuilder } from './util/multi-part-builder';
-import { HttpClient, HttpHeaders, HttpParams, HttpRequest } from '@angular/common/http';
-import { SnackService } from '../../core/snack/snack.service';
-import { SnackType } from '../../core/snack/snack.model';
-import { ConfigService } from '../config/config.service';
-import { GoogleSession } from '../config/config.model';
-import { catchError, concatMap, filter, map, shareReplay, switchMap, take } from 'rxjs/operators';
-import { combineLatest, EMPTY, from, merge, Observable, of, throwError, timer } from 'rxjs';
+import {IS_ELECTRON} from '../../app.constants';
+import {MultiPartBuilder} from './util/multi-part-builder';
+import {HttpClient, HttpHeaders, HttpParams, HttpRequest} from '@angular/common/http';
+import {SnackService} from '../../core/snack/snack.service';
+import {SnackType} from '../../core/snack/snack.model';
+import {ConfigService} from '../config/config.service';
+import {GoogleSession} from '../config/config.model';
+import {catchError, concatMap, filter, map, shareReplay, switchMap, take} from 'rxjs/operators';
+import {combineLatest, EMPTY, from, merge, Observable, of, throwError, timer} from 'rxjs';
 import {
   IPC_GOOGLE_AUTH_TOKEN,
   IPC_GOOGLE_AUTH_TOKEN_ERROR,
   IPC_TRIGGER_GOOGLE_AUTH
 } from '../../../../electron/ipc-events.const';
-import { ElectronService } from 'ngx-electron';
-import { BannerService } from '../../core/banner/banner.service';
+import {ElectronService} from 'ngx-electron';
+import {BannerService} from '../../core/banner/banner.service';
 
 const EXPIRES_SAFETY_MARGIN = 5 * 60 * 1000;
 
@@ -236,6 +236,7 @@ export class GoogleApiService {
     });
   }
 
+  // NOTE: file will always be returned as text (makes sense)
   loadFile(fileId): Observable<any> {
     if (!fileId) {
       this._snackIt('ERROR', 'GoogleApi: No file id specified');
@@ -248,8 +249,9 @@ export class GoogleApiService {
       params: {
         'key': GOOGLE_SETTINGS.API_KEY,
         supportsTeamDrives: true,
-        alt: 'media'
+        alt: 'media',
       },
+      responseType: 'text',
     });
     const metaData = this.getFileInfo(fileId);
 
@@ -264,7 +266,7 @@ export class GoogleApiService {
       );
   }
 
-  saveFile(content, metadata: any = {}): Observable<{}> {
+  saveFile(content: any, metadata: any = {}): Observable<{}> {
     if ((typeof content !== 'string')) {
       content = JSON.stringify(content);
     }
@@ -414,19 +416,21 @@ export class GoogleApiService {
     return from(loginObs)
       .pipe(
         concatMap(() => {
-          const p = {
+          const p: any = {
             ...params_,
             headers: {
               ...(params_.headers || {}),
               'Authorization': `Bearer ${this._session.accessToken}`,
             }
           };
+
           const bodyArg = p.data ? [p.data] : [];
           const allArgs = [...bodyArg, {
             headers: new HttpHeaders(p.headers),
             params: new HttpParams({fromObject: p.params}),
             reportProgress: false,
             observe: 'response',
+            responseType: params_.responseType,
           }];
           const req = new HttpRequest(p.method, p.url, ...allArgs);
           return this._http.request(req);